Use JavaScript to generate header permalinks
Using this approach, we use the built-in ability of Redcarpet to tag header
elements with id values. Then we add a bit of JavaScript to the template for
posts that adds permalinks pointing at these ids.

TODO: The typographer in me wants these links to hang in the margin. Figure out
how to do that with CSS.