Fix small caps font face
Some time in the past 5 years, `font-variant: small-caps` was no longer
enough to cause browsers to recognize a separate sub-category of
`font-family: TeXGyrePagella`. Since this variant is listed last, it
also became the "default" font face. This commit fixes the glitch by
re-naming the font to `font-family: TeXGyrePagella`.

The style rule for `abbr` is updated and also has `text-transform: none`
added as the default for some browsers is to uppercase everything.